+++ GANDLF/compute/training_loop.py
@@ -31,7 +31,7 @@ from GANDLF.utils import (
 from GANDLF.logger import Logger
 from .step import step
 from .forward_pass import validate_network
-from . import create_pytorch_objects
+from .generic import create_pytorch_objects
 
 # hides torchio citation request, see https://github.com/fepegar/torchio/issues/235
 os.environ["TORCHIO_HIDE_CITATION_PROMPT"] = "1"
